3.3.16

+++ Source/Console/HPSDR/JanusAudioImports.cs
@@ -254,10 +254,10 @@ namespace PowerSDR
         public static extern int getFwdPower();
 
         [DllImport("JanusAudio.dll", CallingConvention = CallingConvention.Cdecl)]
-        public static extern int getRefPower();
+        public static extern float getRefPower();
 
         [DllImport("JanusAudio.dll", CallingConvention = CallingConvention.Cdecl)]
-        public static extern int getAlexFwdPower();
+        public static extern float getAlexFwdPower();
 
         [DllImport("JanusAudio.dll", CallingConvention = CallingConvention.Cdecl)]
         public static extern int getAIN3();
